Re-organize the code and optimize a little bit
This change moves all the logic in the combo-breaker module. The visibility
of some APIs is still wrong, but that will be fixed on the @Composable
function meant as the public API is created inside that module.

This change also reduces allocations in several places by avoiding the
nice for-each loops Kotlin provides (hurts readability but kind of
unavoidable, the difference isn't worth the potential cost on older
API levels).

Finally, this change also refactors code a little bit to reuse what
can be reused in the line breaking algorithm. Actual line breaking
is still done in the demo module, but that will also eventually
move to the "business logic" module.