Fix overlapping shapes and several slotting bugs
The existing algorithm was much simpler but couldn't handle shapes to be
defined in arbitrary order. They had to be defined from more or less left
to right which wasn't user friendly at all. Supporting overlapping shapes
is more of a defensive technique than something that should actually be
useful but eh…

With this new implementation, a left or right only flowing shape acts as
a barrier for text flow: nothing will ever be laid out to their respective
right or left.