Add support for multiple shapes per elements
This change introduces the new flowShapes() modifiers to return a list
of Path. It also adds a new Path.divide() extension to create a list
of paths from a single source Path.

This change also fixes an issue with the ordering of text segments
when multiple shapes are on the same line of text.